requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
bcrypt==4.2.0
billiard==4.2.0
celery==5.4.0
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.6.0
cryptography==43.0.0
Django==5.0.7
djangorestframework==3.15.2
dnspython==2.6.1
drf-spectacular==0.27.2
gevent==24.2.1
greenlet==3.0.3
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.3.7
passlib==1.7.4
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pycparser==2.22
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.8
referencing==0.35.1
rpds-py==0.19.1
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
zope.event==5.0
zope.interface==6.4.post2
